Perform linearization after conversion to `ExpEmbedding` (#180)
Prior to this change, `StmtConverter` was responsible for taking the FIR
body of a function and turning it into a Viper `Stmt.Seqn`. This change
splits that process into two steps: we first convert the FIR into a
single `ExpEmbedding` and then we perform linearization on the
`ExpEmbedding` to obtain a `Stmt.Seqn`.

The goal of this PR is to perform the majority of the work moving from
our old approach to our new approach, in order to allow further
development to happen in the new approach. There is still a lot of
clean-up work to be done: I've filed issues about them.