fixup! [K1] Fix false negative DEPRECATED_RESOLVE_WITH_AMBIGUOUS_ENUM_ENTRY
Change the API. It looks like trying to
find a single classifier is OK in some
cases: for them there are the
`getContributedClassifier(?!s)*`
functions.

In some cases it seems that rewriting
the code to account for multiple
classifiers may be the correct thing:
in these cases an explicit
`.singleOrNull()` is used, so that they
(1) stand out in the commit/review
(2) stand out in the code by looking
weird.

Note that this commit should preserve
the semantics. Note that it is ok to take
into account multiple candidates if we
just delegate the call to another scope.

Note that the single-version-functions need
to be extensions instead of members. This is
crucial for scopes that do `: LexicalScope
by workerScope {}`, because otherwise these
functions get delegated which leads to incorrect
behavior in cases like
`DeprecatedLexicalScope`: this scope
needs to wrap `workerScope` results,
meaning we shouldn't delegate
`*IncludeDeprecated()` as is.

See `DiagnosticTestGenerated.testNoMembers_before`.