K2 AA: Render shortened type as string if already imported
KtReferenceShortener is a powerful API used by many components of IJ
like inspections, intentions, quick-fixes, and so on. The reference
shortener has two parts: AA and updating PSI. Unfortunately, it is
not easy to use, when we have to run AA before updating PSI but we
clearly know how to shorten it AFTER updating the PSI.

For example, create-function-from-usage runs AA to collect
information to build a function definition. After we collecting the
information, we create the PSI for the function definition, and we add
it to a container (class or file). However, the reference shortener does
not know the context of the container like what types are already
imported until we actually add created PSI (update the PSI). Since we
cannot run AA on EDT (because of UI freeze), there is no way to run the
analysis part of the reference shortener. Moreover, if we run the
reference shortener early and keep shortening information as
ShortenCommand, the information will be invalid after we update the PSI.

To overcome the drawback of reference shortener, this commit introduces
a new AA:
```
 public fun renderShortenedTypeIfAlreadyImported(
   classId: ClassId, elementForScope: KtElement
 ): String
```
It will return a string form of shortened type if it is already imported
or a part of the source file.

For example,
```
package com.x.y.z
import dependency.Foo
class Bar { class Inner { class MoreInner } }
```
If we call this API with `classId = com.x.y.z.Bar.Inner` and
`elementForScope = KtClass For MoreInner`, it will return "Inner". If we
call this API with `dependency.Foo` and
`elementForScope = KtClass For Bar`, it will return "Foo".
If we call this API with `classId = not.already.imported.TypeX` and
`elementForScope = KtClass For Bar`, it will return
"not.already.imported.TypeX".

This API will be useful for the case we need to shorten types before
we have the final PSI, because updating PSI will make `ShortenCommand`
invalid, which means `collectPossibleReferenceShorteningsInElement` and
`collectPossibleReferenceShortenings` cannot help if the PSI is not
final. Therefore, one of major users of this API will be the
create-from-usage quick-fixes as they need to render lots of types before
it finally updates PSI.