Fix crash during overridden composable types lookup
Some annotations defined in current module can be in partially incomplete state (either completely unbound or parent is missing). When we lookup overriden functions to potentially update them, it results in a crash.

Practically, we only need to update functions in a different module, as deep copy already changes functions in the current one, so this change does exactly that.

Relnote: "Fixes a crash during lookup of overridden composable types"
Test: Compiler transform tests
Fixes: 297665426 ( https://issuetracker.google.com/issues/297665426 )
Change-Id: Ib6d2c0ec3e18716f345a61b4df465cb64d268fec ( https://android-review.git.corp.google.com/q/Ib6d2c0ec3e18716f345a61b4df465cb64d268fec )