Allow removing groups that are no longer required
Added an option, "nonSkippingGroupOptimization" (disabled by default),
that allows removing groups around non-skipping composable functions
(including intrnsic remember) when it is supported by the runtime

The runtime was modified to allow calling `cached` after a child
group was added to the slot table. This means that static calls
that do not skip are no longer required to have a group. This
change modifies the plugin to not generate the groups that are
no longer necessary after the runtime change.

Relnote: """
Added an experimental optimization that will remove the groups
around functions that are not skippable such as explicitly marked
as `@NonSkippableComposable` and functions that are implicitly
not skippable such inline functions and functions that return a
non-Unit value such as `remember`.

This optimization can be enabled by passing plugin option,

  -P plugin:androidx.compose.compiler.plugins.kotlin:nonSkippingGroupOptimization=true

to the Kotlin compiler.
"""

Bug: 254480106 ( https://issuetracker.google.com/issues/254480106 )
Test: ./gradlew :compose:r:r:tDUT :compose:compiler:c-h:i-t:tDUT
Change-Id: I1688f5dbf178696231c9c375c1299cd4a45a9fcd ( https://android-review.git.corp.google.com/q/I1688f5dbf178696231c9c375c1299cd4a45a9fcd )