[Analysis API] Fix overridden declaration computation for fake overrides
According to the commit history, the previous behavior took place
likely because of issues with implementing the override/implement
dialog. However, it broke FIR implementations of
'getAllOverriddenSymbols()' and 'getDirectlyOverriddenSymbols()' in case
if a passed callable is a declaration-side substitution override.

Override/implement works as usual after the change, and all relevant
tests pass.

'originalContainingClassForOverride' that was used before, contains the
original class for substitution overrides (the class from where the
callable was copied), instead of the owner one. 'getContainingSymbol()'
has more consistent behavior.

Note that in the current implementation, fake overrides are always
unwrapped according to the KDoc of API methods.

^KT-62405 Fixed