Give an option to prepare IrClass for IrGenerationExtension
Motivation: Compose compiler plugin adds an annotation to specific
classes even when they do not have the annotation in their source code.
If this is done by a new codegen API KtFirCompilerFacility, the added
annotation causes FIR resolution issues. It happens because the
IrGenerationExtension in the Compose compiler plugin uses the annotation
before FIR resolutation and FIR2IR conversion.

This commit introduces `prepareIrClassUsedInsideExtension` function in
IrGenerationExtension interface to allow its child extension to prepare
IrClassSymbol before using it inside the extension. The
`prepareIrClassUsedInsideExtension` function is called by
`KtFirCompilerFacility.compile()` before `generate()` in the extension
is called.