K2: pass FirFile to FirElementSerializer.classProto
This is based on a previous attempt at 951e30b683, which was previously
reverted in d83f6a40fb, but now is needed again to implement KT-57919,
because in order to use the IR-based constant expression evaluator and
not fall back to the FIR-based one, serialization of annotation
arguments on K2 requires to know the containing file (see
`ConstValueProvider.processingFirFile`). Specifically, see the test
`evaluateConstructorOfUnsignedType.kt`, where FIR-based evaluator fails
(as expected) because unsigned type constructors are invisible.

 #KT-57919