[LLVM 16] Ignore more symbols when linking with `libclangext`
Currently, when linking `libclangstubs` with `libclangext` included,
we have to add a number of LLVM/Clang components as well.
To avoid adding unnecessary parts, we let some unused symbols be
undefined - namely, instruct the linker to ignore the fact that some
symbols are undefined, which is fine if they are not used during
execution.

This commit just expands the list of such symbols as a preparation for
the update to LLVM 16.

^KT-49279