[K/N][tests] Use LIBCLANG_DISABLE_CRASH_RECOVERY in the "new" test infra
This environment variable is also used for the cinterop tool itself.
So it is good to have it in the tests as well, just to make the tests
match the production as closely as possible.

This variable alters the way libclang handles signals, so its effect is
non-trivial and worth keeping in sync.